# asset-log/docs/openapi.json
{
  "components": {
    "schemas": {
...
      "CreateTransaction": {
        "properties": {
          "account_id": {
            "format": "uuid",
            "type": "string"
          },
          "asset_id": {
            "format": "uuid",
            "type": "string"
          },
          "fee": {
            "description": "手数料。0以上。未指定なら0",
            "example": "550",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/TradeKind"
          },
          "note": {
            "type": [
              "string",
              "null"
            ]
          },
          "price": {
            "description": "単価。0以上",
            "example": "2350.5",
            "type": "string"
          },
          "quantity": {
            "description": "数量。正の数",
            "example": "100",
            "type": "string"
          },
          "traded_at": {
            "description": "約定日。未来日は不可（日本時間で判定）",
            "format": "date",
            "type": "string"
          }
        },
        "required": [
          "account_id",
          "asset_id",
          "kind",
          "quantity",
          "price",
          "traded_at"
        ],
        "type": "object"
      },
...
      "FieldError": {
        "properties": {
          "field": {
            "description": "エラーが発生した入力項目名",
            "example": "quantity",
            "type": "string"
          },
          "message": {
            "description": "項目ごとのエラー内容",
            "example": "数量は正の数である必要があります",
            "type": "string"
          }
        },
        "required": [
          "field",
          "message"
        ],
        "type": "object"
      },
...
      "ProblemDetails": {
        "description": "RFC 9457 準拠のエラーレスポンス（`application/problem+json`）",
        "properties": {
          "detail": {
            "description": "人間向けの詳細メッセージ",
            "example": "口座が見つかりません",
            "type": "string"
          },
          "errors": {
            "description": "項目単位のエラー（422のみ。空の場合はフィールドごと省略される）",
            "items": {
              "$ref": "#/components/schemas/FieldError"
            },
            "type": "array"
          },
          "status": {
            "description": "HTTPステータスコード",
            "example": 404,
            "format": "int32",
            "maximum": 599,
            "minimum": 100,
            "type": "integer"
          },
          "title": {
            "description": "エラー種別の短い説明",
            "example": "Not Found",
            "type": "string"
          },
          "trace_id": {
            "description": "サーバーログと突き合わせるための識別子",
            "format": "uuid",
            "type": "string"
          },
          "type": {
            "description": "エラー種別の安定した識別子。クライアントはこの値で分岐できる",
            "example": "/errors/not-found",
            "type": "string"
          }
        },
        "required": [
          "type",
          "title",
          "status",
          "detail",
          "trace_id"
        ],
        "title": "ProblemDetails",
        "type": "object"
      },
...
      "TradeKind": {
        "description": "取引種別。",
        "enum": [
          "buy",
          "sell"
        ],
        "type": "string"
      },
      "TransactionResponse": {
        "properties": {
          "account_id": {
            "format": "uuid",
            "type": "string"
          },
          "asset_id": {
            "format": "uuid",
            "type": "string"
          },
          "created_at": {
            "format": "date-time",
            "type": "string"
          },
          "fee": {
            "example": "550",
            "type": "string"
          },
          "id": {
            "format": "uuid",
            "type": "string"
          },
          "kind": {
            "$ref": "#/components/schemas/TradeKind"
          },
          "note": {
            "type": [
              "string",
              "null"
            ]
          },
          "price": {
            "example": "2350.5",
            "type": "string"
          },
          "quantity": {
            "example": "100",
            "type": "string"
          },
          "traded_at": {
            "format": "date",
            "type": "string"
          }
        },
        "required": [
          "id",
          "account_id",
          "asset_id",
          "kind",
          "quantity",
          "price",
          "fee",
          "traded_at",
          "created_at"
        ],
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      },
      "jobToken": {
        "description": "バッチ実行用の固定トークン",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "NISA/iDeCo を含む資産・取引の管理API",
    "license": {
      "name": ""
    },
    "title": "asset-log API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/transactions": {
      "get": {
        "operationId": "list_transactions",
        "parameters": [
          {
            "description": "口座で絞り込む",
            "in": "query",
            "name": "account_id",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "description": "銘柄で絞り込む",
            "in": "query",
            "name": "asset_id",
            "required": false,
            "schema": {
              "format": "uuid",
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "description": "約定日の開始（含む）",
            "in": "query",
            "name": "from",
            "required": false,
            "schema": {
              "format": "date",
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "description": "約定日の終了（含む）",
            "in": "query",
            "name": "to",
            "required": false,
            "schema": {
              "format": "date",
              "type": [
                "string",
                "null"
              ]
            }
          },
          {
            "description": "取得件数。既定100、最大500に丸められる",
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "format": "int64",
              "type": [
                "integer",
                "null"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TransactionResponse"
                  },
                  "type": "array"
                }
              }
            },
            "description": "取引の一覧（約定日の降順）"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "認証が必要"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "開始日が終了日より後"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "transactions"
        ]
      },
      "post": {
        "operationId": "create_transaction",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTransaction"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            },
            "description": "取引を登録した"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "認証が必要"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "口座または銘柄が存在しない"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "数量が0以下、価格や手数料が負、未来日、売却が保有数量を超える"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "transactions"
        ]
      }
    },
    "/transactions/{id}": {
      "delete": {
        "operationId": "delete_transaction",
        "parameters": [
          {
            "description": "取引ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "削除した"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "認証が必要"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "取引または銘柄が存在しない"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "削除すると以降の売却が保有数量を超える"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "transactions"
        ]
      },
      "get": {
        "operationId": "get_transaction",
        "parameters": [
          {
            "description": "取引ID",
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "format": "uuid",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransactionResponse"
                }
              }
            },
            "description": "取引の詳細"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "認証が必要"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            },
            "description": "取引が存在しない"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "tags": [
          "transactions"
        ]
      }
    }
  },
  "servers": [
    {
      "description": "ローカル開発",
      "url": "http://localhost:8080"
    }
  ],
...
}